Change the semantics of start/stop and auto enabling.
Remove auto-detaching if no markers are present anymore.
Instead control when the start event is triggered.

At startup things are not explicitly enabled or disabled.
If auto enabling is set, the start event will be triggered in this case.

After an explicit call to start/stop things behave as you would expect.